refactoring code to use ES6 standards - midway through
