+++ notes.js
@@ -1,0 +1,132 @@
+/*
+Code that isn't used but will be useful for reference later
+ */
+
+
+
+
+
+
+
+
+/*
+Facebook calls to get user info and picture
+ */
+
+
+
+
+var Photo = React.createClass({
+  propTypes: {
+    user: React.PropTypes.object.isRequired,
+  },
+
+  getInitialState: function(){
+    return {
+      photo: null,
+    };
+  },
+
+  componentWillMount: function(){
+    var user = this.props.user;
+    var api = `https://graph.facebook.com/v2.3/${user.userId}/picture?width=${FB_PHOTO_WIDTH}&redirect=false&access_token=${user.token}`;
+    var self = this;
+
+    fetch(api)
+      .then((response) => response.json())
+      .then((responseData) => {
+        self.setState({
+          photo : {
+            url : responseData.data.url,
+            height: responseData.data.height,
+            width: responseData.data.width,
+          }
+        });
+      }).done();
+  },
+
+  render: function(){
+    if(this.state.photo == null) return this.renderLoading();
+    
+    var photo = this.state.photo;
+
+    return (
+      <View style={styles.bottomBump}>
+
+        <Image
+          style={photo &&
+            {
+              height: photo.height,
+              width: photo.width,
+            }
+          }
+          source={{uri: photo && photo.url}}
+        />
+      </View>
+    );
+  },
+  renderLoading: function(){
+    return (
+      <View>
+        <Text>Loading...</Text>
+      </View>
+    );
+  }
+});
+
+var Info = React.createClass({
+  propTypes: {
+    user: React.PropTypes.object.isRequired,
+  },
+
+  getInitialState: function(){
+    return {
+      info: null,
+    };
+  },
+
+  componentWillMount: function(){
+    var _this = this;
+    var user = this.props.user;
+    var api = `https://graph.facebook.com/v2.3/${user.userId}?fields=name,email&access_token=${user.token}`;
+
+    fetch(api)
+      .then((response) => response.json())
+      .then((responseData) => {
+        _this.setState({
+          info : {
+            name : responseData.name,
+            email: responseData.email,
+          },
+        });
+      })
+      .done();
+  },
+
+  render: function(){
+    var info = this.state.info;
+
+    return (
+      <View style={styles.bottomBump}>
+        <Text>{ info && this.props.user.userId }</Text>
+        <Text>{ info && info.name }</Text>
+        <Text>{ info && info.email }</Text>
+      </View>
+    );
+  }
+});
+
+var styles = StyleSheet.create({
+  loginContainer: {
+    marginTop: 150,
+
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  bottomBump: {
+    marginBottom: 15,
+  },
+});
+
+
