--- src/App.js
@@ -1,21 +1,0 @@
-var React = require('react-native');
-var { View, StyleSheet, Text, Navigator } = React;
-var SplashPage = require('./SplashPage');
-var LoginPage = require('./LoginPage');
-var TrailMap = require('./TrailMap');
-
-module.exports = React.createClass({
-
- render: function() {
-    return ( "");
-  } 
-});
-
-var styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  }
-});
