+++ src/main.js
@@ -1,10 +1,15 @@
-import React, {
+'use strict';
+
+const React = require('react-native');
+const {
+  Component,
   StyleSheet,
-  Navigator,
-} from 'react-native';
-import SplashPage from './SplashPage';
-import LoginPage from './LoginPage';
-import TrailMap from './TrailMap';
+  Navigator
+} = React;
+
+const SplashPage = require('./SplashPage');
+const LoginPage = require('./LoginPage');
+const TrailMap = require('./TrailMap');
 
 const ROUTES = {
   splash: SplashPage,
@@ -12,25 +17,36 @@ const ROUTES = {
   trailmap: TrailMap
 };
 
-module.exports = React.createClass({
-  renderScene: function(route, navigator) {
-    var Component = ROUTES[route.name]; //ROUTES['signin'] => Signin
-    return <Component route={route} navigator={navigator} />;
-  },
-  render: function(){
-    return (
-      <Navigator
-        style={styles.container}
-        initialRoute={{name: 'splash'}}
-        renderScene={this.renderScene}
-        configureScene={()=> { return Navigator.SceneConfigs.FloatFromRight; }}
-        />
-    )
-  }
-});
+class Main extends Component {
 
-var styles = StyleSheet.create({
+  constructor(props) {
+
+      super(props);
+  }
+
+  renderScene(route, navigator) {
+    console.log('Route to ' + route.name);
+      var Component = ROUTES[route.name]; //ROUTES['signin'] => Signin
+      return <Component route={route} navigator={navigator} />;
+  }
+
+  render(){
+
+      return (
+        <Navigator
+          style={styles.container}
+          initialRoute={{name: 'splash'}}
+          renderScene={this.renderScene}
+          configureScene={()=> { return Navigator.SceneConfigs.FloatFromRight; }}
+          />
+      );
+    }
+}
+
+const styles = StyleSheet.create({
   container: {
     flex: 1
   }
 });
+
+module.exports = Main;
